feat: add stepwise process for shared request
