+++ packages/hoppscotch-common/src/components.d.ts
@@ -186,7 +186,9 @@ declare module 'vue' {
     SettingsExtension: typeof import('./components/settings/Extension.vue')['default']
     SettingsProxy: typeof import('./components/settings/Proxy.vue')['default']
     Share: typeof import('./components/share/index.vue')['default']
+    ShareCreateModal: typeof import('./components/share/CreateModal.vue')['default']
     ShareCustomizeModal: typeof import('./components/share/CustomizeModal.vue')['default']
+    ShareModal: typeof import('./components/share/Modal.vue')['default']
     ShareRequest: typeof import('./components/share/Request.vue')['default']
     ShareRequestModal: typeof import('./components/share/RequestModal.vue')['default']
     ShareShareRequestModal: typeof import('./components/share/ShareRequestModal.vue')['default']
